requirements.txt
Adafruit-GPIO==1.0.3
appdirs==1.4.4
CacheControl==0.12.6
cffi==1.14.6
chardet==4.0.0
click==8.0.3
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==35.0.0
distlib==0.3.3
distro==1.6.0
Flask==2.0.2
Flask-Cors==3.0.10
html5lib==1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
more-itertools==8.10.0
msgpack==1.0.2
ordered-set==4.0.2
packaging==20.9
pep517==0.11.0
ply==3.11
progress==1.6
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
RPi.GPIO==0.6.3
six==1.16.0
toml==0.10.2
tomli==1.2.1
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2